Esqueci a senha

+++ src/main/java/com/alves/service/validation/ClienteUpdateValidator.java
@@ -34,7 +34,7 @@ public class ClienteUpdateValidator implements ConstraintValidator<ClienteUpdate
 		
 		@SuppressWarnings("unchecked")
 		Map<String, String> map = (Map<String, String>) request.getAttribute(HandlerMapping.URI_TEMPLATE_VARIABLES_ATTRIBUTE);
-		Long uriId = Long.parseLong(map.get("id"));
+		Integer uriId = Integer.parseInt(map.get("id"));
 		
 		List<FieldMessage> list = new ArrayList<>();
